pyproject.toml
[tool.poetry]
name = "liqui_hist"
version = "0.1.0"
description = ""
authors = ["LazyGeek"]

[tool.poetry.dependencies]
python = "^3.12"
pytz = "^2024.2"
websockets = "^13.0.1"
termcolor = "^2.4.0"
requests = "^2.31.0"
mysql-connector-python = "^8.0.33"
fastapi = "^0.114.1"

[data.dependencies.package-mode]
package-mode = false

[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true
